Automatically create tracker events from event request forms.
Still has a ways to go, creates an event a year from now with most of the
details in the notes field. Requires adding tracker login and password to
secrets file. Also, I am not proud of this code, but I am proud of what it
will accomplish....if it works.